Add receipt endpoint
Adds a GET receipt endpoint which generates a receipt for a transaction (izettle_transactions only).

Receipt is generated using rocket_dyn_templates with handlebars thus adding a new dependency for the backend.

Also changes the type of the name field for InventoryItem from String to Option<String> as this appears to be the value in the database (presumably this type was a bug?).